requirements.txt
langchain
langchain_openai
python-dotenv
langchain_community
tweepy